No longer terminating when received illegal auxtemp frame Camera now prints string error instead of int error from libgphoto2 Fixed bug in File resource regarding still_id input checking Corrected types of probe and endstop values in Frame to match those in DB Removed commented code in HTTPResponseBody Fixed bug in calculation of sleep time; added logging when cycle cannot keep up with desired interva;

+++ prosurd/src/Database/Frame.hpp
@@ -29,8 +29,8 @@ namespace Prosur::Database{
 		vector<float> cpuTemp;
 		vector<float> inputVoltage;
 		vector<float> motorPos; // May be used for x, y, z, e for now. Flexible.
-		vector<float> endstop; // Suggested use: x-endstop, y-endstop, z-probe, filament runout sensor (x,y,z,e)
-		vector<float> probe;
+		vector<int> endstop; // Suggested use: x-endstop, y-endstop, z-probe, filament runout sensor (x,y,z,e)
+		vector<int> probe;
 
 		float speedCurrentMms;
 		float speedRequestedMms;
